requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
pymysql

python-dotenv
passlib[bcrypt]
python-jose[cryptography]

pydantic
email-validator

tensorflow
numpy
opencv-python-headless

groq
python-multipart
